Fixed teleportation through solid objects
Moved teleport implementation into a dedicated function.

Removed the ability to teleport through or into map geometry.
When trying to teleport, the function checks for objects in front of the player through a linetrace. When nothing is hit by the raycast, the player is teleported 10 meters forward. If there is an object in front, the player is teleported as close to the hitpoint as possible (not on the exact location of the hitpoint, to avoid partially clipping through the object).