Jetpack ability first implementation
First implementation of the Jetpack ability. No visual or sound effects are present at this stage. No relative UI elements have been implemented either.

By pressing the jump button while mid-air, the player is able to briefly fly, mimicking the functioning of a jetpack, until the energy pool is depleted.

This is achieved by increasing the velocity of the custom character movement component on the Z axis. The jetpack force is increased at every frame up to a limit, taking into consideration the variable time delta between frames. DeltaTime is passed to the character movement from the ShooterCharacter Tick function at every frame. Special care needed to be taken when thinking about multiplayer functioning: frame delta times between client and server could differ in a significant way, and for this reason the server's DeltaTime variable is set through an RPC (client sends the server its value of the variable to avoid differences in the jetpack force and the subsequent server corrections which would cause a rubber-band effect).

- ShooterCharacter handles the energy pool depletion and recharge, controls the ability to actually activate the jetpack when pressing the input and sends the deltatime to the ShooterMovement.

- ShooterCharacterMovement is where the jetpack ability and networking is actually implemented