Improved jetpack functioning and added UI
Activating the jetpack now sets the movement mode to "Flying", so gravity is not in effect when in this movement mode. This solved a bug where the behaviour of the jetpack changed depending on the framerate (even though DeltaSeconds was in use to calculate vertical velocity), as if the gravity increased with the decrease of framerate. The player also has better control over the movement while in flyigng mode.

Added a simple jetpack energy bar in the UI using a widget, with a mix of C++ and blueprints.