Wall jump ability implementation
First implementation of the wall jump ability.

By pressing the jump button while next to a (somewhat vertical)  solid surface, the player is able to perform a jump off the wall.

When jumping while being airborne, added a collision check is performed to detect walls near the player. If one is detected, the normal to the hit surface is saved, and, if its angle is suitable, the wall jump ability is triggered. When calling WallJumpPressed, the normal vector is also sent to the CharacterMovementComponent.

The ability is implemented by setting the character's velocity on the Z axis equal to the JumpZVelocity (giving the behaviour of a regular jump), and summing the X and Y components of the character's velocitiy with the respective values of the normal vector (giving the lateral push away from the wall).

In order to implement client prediction and server reconciliation, other than the same techniques used for the previous abilities, the normal vector also needed to be transfered via a server RPC.